refactor: move override parsing into access package, fix API key enforcement
Address PR review feedback:

- Move ScopeOverride type and parseOverrideHeader into access/override.go
  so the access package doesn't import middleware (dependency inversion fix)
- Middleware now stores raw header string; access package parses it
- Skip override enforcement for API key requests so the dev toolbar
  doesn't interfere with API key auth flows
- Use early continue in grantsFromOverrides per style suggestion
